+++ ali/fabric/src/main/java/com/yanny/ali/fabric/platform/FabricPlatformHelper.java
@@ -1,5 +1,6 @@
 package com.yanny.ali.fabric.platform;
 
+import com.google.common.base.Suppliers;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.yanny.aci.CommonLogUtils;
@@ -14,14 +15,18 @@ import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.item.SpawnEggItem;
 import net.minecraft.world.level.storage.loot.LootPool;
 import net.minecraft.world.level.storage.loot.LootTable;
+import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 
 import java.nio.file.Path;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.function.Supplier;
 
 public class FabricPlatformHelper implements IPlatformHelper {
     private static final Logger LOGGER = CommonLogUtils.getLogger(Utils.MOD_ID);
+
+    private final Supplier<List<IPlugin>> pluginsSupplier = Suppliers.memoize(this::loadPlugins);
 
     @Override
     public List<LootPool> getLootPools(LootTable table) {
@@ -30,6 +35,11 @@ public class FabricPlatformHelper implements IPlatformHelper {
 
     @Override
     public List<IPlugin> getPlugins() {
+        return pluginsSupplier.get();
+    }
+
+    @NotNull
+    private List<IPlugin> loadPlugins() {
         List<IPlugin> plugins = new LinkedList<>();
 
         for (EntrypointContainer<IPlugin> container : FabricLoader.getInstance().getEntrypointContainers("ali", IPlugin.class)) {