+++ alicompat/common/src/main/java/com/yanny/alicompat/IGlmModCompat.java
@@ -1,0 +1,8 @@
+package com.yanny.alicompat;
+
+import com.yanny.ali.plugin.glm.IGlobalLootModifierPlugin;
+import com.yanny.ali.plugin.glm.ILootTableIdConditionPredicate;
+
+public interface IGlmModCompat extends IModCompat {
+    void registerGlobalLootModifier(IGlobalLootModifierPlugin.IRegistry registry, ILootTableIdConditionPredicate predicate);
+}