+++ alicompat/common/src/main/java/com/yanny/alicompat/Utils.java
@@ -1,0 +1,5 @@
+package com.yanny.alicompat;
+
+public class Utils {
+    public static final String MOD_ID = "alicompat";
+}