+++ alicompat/common/src/main/java/com/yanny/alicompat/platform/ICompatPlatform.java
@@ -1,0 +1,7 @@
+package com.yanny.alicompat.platform;
+
+import org.jetbrains.annotations.NotNull;
+
+public interface ICompatPlatform {
+    boolean isModLoaded(@NotNull String modId);
+}