+++ alicompat/fabric/src/compat/farmersdelight/java/com/yanny/alicompat/compat/farmersdelight/package-info.java
@@ -1,0 +1,5 @@
+@ParametersAreNonnullByDefault
+package com.yanny.alicompat.compat.farmersdelight;
+
+
+import javax.annotation.ParametersAreNonnullByDefault;