+++ alicompat/forge/src/compat/farmersdelight/java/com/yanny/alicompat/compat/farmersdelight/CopySkilletFunctionAccessor.java
@@ -1,0 +1,22 @@
+package com.yanny.alicompat.compat.farmersdelight;
+
+import com.yanny.aci.tooltip.TooltipBuilder;
+import com.yanny.ali.api.IServerUtils;
+import com.yanny.ali.language.Lang;
+import com.yanny.ali.plugin.mods.ConditionalFunction;
+import com.yanny.ali.plugin.mods.IFunctionTooltip;
+import net.minecraft.world.level.storage.loot.functions.LootItemConditionalFunction;
+
+public class CopySkilletFunctionAccessor extends ConditionalFunction implements IFunctionTooltip {
+    public CopySkilletFunctionAccessor(LootItemConditionalFunction conditionalFunction) {
+        super(conditionalFunction);
+    }
+
+    @Override
+    public TooltipBuilder getTooltip(IServerUtils utils) {
+        return TooltipBuilder.array((b) -> {
+            b.add(utils.getValueTooltip(utils, predicates).build(Lang.Branch.PREDICATES));
+            b.showEmpty();
+        }, FarmersDelightLang.Functions.COPY_SKILLET);
+    }
+}